refactor: isolate Responses request diagnostics
